+++ wpimath/src/main/native/include/wpimath/MathShared.h
